move to eloquent component
